requirements.txt
selenium==4.10.0
webdriver-manager==3.8.6
pytest==7.4.0
pyyaml
undetected_chromedriver